Revert "try out honeycomb frontend monitoring (#217)"
This reverts commit 75b3607fe56a6148eb6f0776e225baf18955c035.
